remove loadintercom
